+++ drl/algos/common/__init__.py
@@ -1,23 +1,22 @@
+from drl.algos.common.credit_assignment import (
+    extract_reward_name, get_credit_assignment_ops, GAE, NStepAdvantageEstimator
+)
+from drl.algos.common.grad_ops import apply_pcgrad
+from drl.algos.common.losses import get_loss
 from drl.algos.common.metrics import (
     global_mean, global_means, global_gather, global_gathers,
     MultiDeque, pretty_print
 )
+from drl.algos.common.schedules import LinearSchedule
 from drl.algos.common.trajectory import TrajectoryManager
-from drl.algos.common.losses import get_loss
 from drl.algos.common.wrapper_ops import update_trainable_wrappers
-from drl.algos.common.grad_ops import apply_pcgrad
-from drl.algos.common.credit_assignment import (
-    extract_reward_name, get_credit_assignment_ops, GAE, NStepAdvantageEstimator
-)
-
 
 __all__ = [
-    "global_mean", "global_means", "global_gather", "global_gathers",
-    "MultiDeque", "pretty_print",
-    "TrajectoryManager",
-    "get_loss",
-    "update_trainable_wrappers",
+    "extract_reward_name", "get_credit_assignment_ops", "GAE", "NStepAdvantageEstimator",
     "apply_pcgrad",
-    "extract_reward_name",
-    "get_credit_assignment_ops", "GAE", "NStepAdvantageEstimator"
+    "get_loss",
+    "global_mean", "global_means", "global_gather", "global_gathers", "MultiDeque", "pretty_print",
+    "LinearSchedule",
+    "TrajectoryManager",
+    "update_trainable_wrappers",
 ]
